docs: reconcile todo.md against verified test/clippy state; fix just-clippy regression
Re-ran the actual verification commands (h264 lib tests, ITU conformance
suite, av1 lib tests, workspace clippy -D warnings, fmt --check) rather
than trusting prior todo.md prose, per CLAUDE.md's "check code before
trusting todo.md checkboxes".

Found `cargo clippy --workspace --all-targets -- -D warnings` was
actually failing (nonminimal_bool in ref_pic.rs's trace_ref_list debug
helper and two scratch CABAC-CBP oracle test files, plus an
unnecessary_cast in another) -- fixed all four, no behavior change.

Updated todo.md's top index: corrected a stale claim that AV1 inter
prediction is unimplemented (the code path has been wired since
2026-08-16; what's actually unverified is bit-exactness, since this
environment has no ffmpeg/dav1d to re-run those gated tests), and added
the 2026-09-05 H.264 (SPS/PPS-by-id fix, MIDR_MW_D frame_num-gap
root-cause, CABAC multi-slice end_of_slice_flag fix) and AV1 (delta_lf
wiring, inter-block loop-filter metadata gap, mandelbrot deblock/CDEF
investigation) session summaries that were committed to
todo-h264.md/todo-av1.md but not yet reflected in the index.

+++ tpt-kinetix-h264/tests/dbg_engine_diff.rs
@@ -434,7 +434,7 @@ fn engine_lockstep_vs_ffmpeg() {
             // Empirically verified (single_step_probe): ffmpeg packs states as
             // 2*pStateIdx + valMPS — identical semantics to the crate's split
             // representation.
-            ff_states[i] = ((crate_ctx[i].state << 1) | (crate_ctx[i].mps & 1)) as u8;
+            ff_states[i] = (crate_ctx[i].state << 1) | (crate_ctx[i].mps & 1);
         }
 
         match run_payload(&payload, &tables, &mut ff_states, &mut crate_ctx, &ops) {
